bitcoinpaperwallet.com is under new ownership
At the end of April 2018, I sold the bitcoinpaperwallet.com website,
service, and associated domain names to a new owner.  This also
includes related projects such as litecoinpaperwallet.com and
ethereumpaperwallet.com.  All orders for hologram stickers and CDs, and
any questions relating to the business, should be directed to the new
owner at orders@paperwalletshop.com.  Also, I will no longer be making
updates or accepting code contributions for this project at
https://github.com/cantonbecker/bitcoinpaperwallet. The updates posted
in September 2017 are my final contributions to this project.